0.3.5 - Fix and create Fireplace

+++ puzzle/src/main/java/com/example/puzzle/GraphicsPuz.java
@@ -54,8 +54,8 @@ public class GraphicsPuz implements GraphicsPuzService {
     }
 
     @Override
-    public void drawTexture(Bitmap textureGame, int x, int y) {
-        canvasGame.drawBitmap(textureGame, x, y, null);
+    public void drawTexture(Bitmap textureGame, int x, double y) {
+        canvasGame.drawBitmap(textureGame, x, (float)y, null);
     }
 
     @Override
